Rename to GetBlobByPath
